requirements.txt
web3==6.20.0
py-solc-x==2.0.3
eth-account==0.11.0
pytest==8.3.3
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
supabase==2.4.6
python-dotenv==1.0.1
